+++ R/embeds_server.R
@@ -75,6 +75,12 @@ embeds_server <- function(input, output, session){
 
       if(rv$chart == "world-map")
         e <- mod_world_echarts(df)
+
+      if(rv$chart == "cases-added")
+        e <- mode_new_cases_echarts(df)
+
+      if(rv$chart == "world-timeline")
+        e <- mod_china_others_echarts(df)
       
     } else if(rv$data == "weixin"){