+++ dev/02_dev.R
@@ -21,6 +21,8 @@ golem::add_module( name = "geo_lines" ) # Name of the module
 golem::add_module( name = "time_provinces" ) # Name of the module
 golem::add_module( name = "jhu_death_rate" ) # Name of the module
 golem::add_module( name = "news" ) # Name of the module
+golem::add_module( name = "new_cases" ) # Name of the module
+golem::add_module( name = "china_others" ) # Name of the module
 
 ## 2.2 Add dependencies